+++ src/main/java/org/example/oddventure/domain/admin/controller/AdminController.java
@@ -82,8 +82,8 @@ public class AdminController {
     @PostMapping("/matches/{matchId}/odds")
     public ResponseEntity<ApiResponse<MatchAdminResponse>> setInitialOdds(
             @PathVariable Long matchId,
-            @Valid @RequestBody InitialOddsSetRequest request)
-    {
+            @Valid @RequestBody InitialOddsSetRequest request
+    ) {
         MatchAdminResponse response = adminService.setInitialOdds(matchId, request);
 
         return ApiResponse.success(response, "초기 배당률 설정에 성공했습니다.");
