+++ src/test/java/org/example/oddventure/domain/match/unit/MatchControllerTest.java
@@ -1,4 +1,4 @@
-package org.example.oddventure.domain.match.controller;
+package org.example.oddventure.domain.match.unit;
 
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
@@ -12,6 +12,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 import org.example.oddventure.domain.auth.config.SecurityConfig;
 import org.example.oddventure.domain.auth.jwt.JwtUtil;
+import org.example.oddventure.domain.match.controller.MatchController;
 import org.example.oddventure.domain.match.dto.request.MatchSearchCondition;
 import org.example.oddventure.domain.match.dto.response.MatchResponse;
 import org.example.oddventure.domain.match.enums.MatchStatus;
